chore: move theme switcher
